Refactor two classes into a spec package
The previous package structure caused a package cycle.

Part of gradle/gradle-native#191
